tests/wpt/tests/speculation-rules/prerender/fetch-intercepted-by-service-worker.https.html

<!DOCTYPE html>
<title>Service worker intercepts a fetch request coming from a prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const uid = token();

  const PAGE_URL = `resources/fetch-intercepted-by-service-worker.html?uid=${uid}`;
  const WORKER_URL = 'resources/fetch-intercept-worker.js';

  const registration =
      await service_worker_unregister_and_register(t, WORKER_URL, PAGE_URL);
  t.add_cleanup(() => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');

  const bc = new PrerenderChannel('prerender-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  startPrerendering(PAGE_URL);

  const result = await gotMessage;
  assert_equals(
    result, 'intercepted by service worker',
    "fetch() should go through a service worker's fetch event handler");
}, 'fetch() in a prerendering page should go through a service worker');
</script>
